requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
jsonschema>=4.0
openapi-spec-validator>=0.7
grpcio-tools>=1.60
networkx>=3.0
astroid>=3.0
pytest>=8.0
pytest-cov>=4.0
rich>=13.0
click>=8.0
pydantic>=2.0
